Feature/role based authorization

+++ TaskManagerAPI/Controllers/TasksController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 using TaskManagerAPI.Entities;
 using TaskManagerAPI.Models.Task;
 using TaskManagerAPI.Repositories.Interfaces;
@@ -7,6 +9,7 @@ namespace TaskManagerAPI.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
+    [Authorize]
     public class TasksController : ControllerBase
     {
         private readonly ITaskRepo _taskRepo;
@@ -19,22 +22,48 @@ namespace TaskManagerAPI.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<TaskItem?>> GetById(Guid id)
         {
-            TaskItem? item = await _taskRepo.GetById(id);
-            if (item == null)
+            if (!Guid.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out Guid userId))
+            {
+                return Unauthorized();
+            }
+            TaskItem? taskItem = await _taskRepo.GetById(id);
+            if (taskItem == null)
             {
                 return NotFound();
             }
-            return item;
+            if (User.IsInRole(UserRole.Admin.ToString()))
+            {
+                return Ok(taskItem);
+            }
+            if (userId != taskItem.UserId)
+            {
+                return Forbid();
+            }
+            return Ok(taskItem);
         }
-        [HttpGet("")]
+        [HttpGet()]
         public async Task<ActionResult<List<TaskItem>>> GetAll()
         {
-            return await _taskRepo.GetAll();
+            if (!Guid.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out Guid userId))
+            {
+                return Unauthorized();
+            }
+            var tasks = await _taskRepo.GetAll();
+
+            if (User.IsInRole(UserRole.Admin.ToString()))
+            {
+                return Ok(tasks);
+            }
+            return Ok(tasks.Where(x => x.UserId == userId));
         }
         [HttpPost]
         public async Task<ActionResult<TaskItem>> Create(CreateTaskRequest request)
         {
-            var newTask = new TaskItem(request.Title, request.Description, request.DueDate);
+            if (!Guid.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out Guid userId))
+            {
+                return Unauthorized();
+            }
+            var newTask = new TaskItem(userId, request.Title, request.Description, request.DueDate);
             await _taskRepo.Insert(newTask);
 
             return CreatedAtAction(nameof(GetById), new { id = newTask.Id }, newTask);
@@ -42,10 +71,18 @@ namespace TaskManagerAPI.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(Guid id, UpdateTaskRequest request)
         {
+            if (!Guid.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out Guid userId))
+            {
+                return Unauthorized();
+            }
             TaskItem? taskItem = await _taskRepo.GetById(id);
             if (taskItem == null)
             {
                 return NotFound();
+            }
+            if (!User.IsInRole(UserRole.Admin.ToString()) && userId != taskItem.UserId)
+            {
+                return Forbid();
             }
             if (!string.IsNullOrWhiteSpace(request.Title))
             {
@@ -66,10 +103,18 @@ namespace TaskManagerAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            if (!Guid.TryParse(User.FindFirstValue(ClaimTypes.NameIdentifier), out Guid userId))
+            {
+                return Unauthorized();
+            }
             TaskItem? taskItem = await _taskRepo.GetById(id);
             if (taskItem == null)
             {
                 return NotFound();
+            }
+            if (!User.IsInRole(UserRole.Admin.ToString()) && userId != taskItem.UserId)
+            {
+                return Forbid();
             }
             await _taskRepo.Delete(id);
             return NoContent();
